+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Added
+- Requests are now automatically retried for (some) failures. The new RetryConfiguration struct controls the exact mechanisms for retrying.
+
+### Removed
+- The NotificationCenter events for request completion have been removed. For globally snooping on requests, use a custom LogDestination.
+
+### Changed
+- The names and payload of some of the request related LogEvents have changed, generally to include more data on requests.
+- LogEvent.requestCompleted is no longer sent for requests that fail based on non 2xx HTTP codes, just LogEvent.requestFailure. You can check for a NetableError.httpError payload if you want to log the status code or raw data.
+- LogDestination is now always called on the main thread
+- Default logging now logs URLs on request compeltion, and logs HTTP status codes for HTTP failures
+- Change RequestIdentifier to be opaque
+
 ## [0.9.1] - 24-06-20
 ### Changed
 - Fixed build on non-Apple platforms (note: only basic compliation has been tested, still might not work properly)